refactor: Move lots of code into components
For better overview still WIP
