+++ frontend/package.json
@@ -9,23 +9,19 @@
     "lint": "next lint"
   },
   "dependencies": {
-    "@supabase/auth-helpers-nextjs": "0.2.5",
-    "@supabase/supabase-auth-helpers": "1.4.2",
+    "@supabase/auth-helpers-nextjs": "0.5.3",
+    "@supabase/auth-helpers-react": "0.3.1",
+    "@supabase/auth-ui-react": "0.2.6",
     "@supabase/ui": "0.36.5",
-    "mapbox-gl": "2.10.0",
-    "next": "12.2.3",
-    "react": "17.0.2",
-    "react-dom": "17.0.2",
-    "react-form": "4.0.1"
+    "mapbox-gl": "2.10.0"
   },
   "devDependencies": {
     "@types/mapbox-gl": "2.7.5",
     "@types/node": "18.6.2",
-    "@types/react": "18.0.15",
-    "@types/react-dom": "18.0.6",
-    "@types/react-form": "4.0.2",
+    "@types/react": "17.0.53",
     "eslint": "8.20.0",
     "eslint-config-next": "12.2.3",
-    "typescript": "4.7.4"
+    "next": "12.3.4",
+    "typescript": "4.9.4"
   }
 }
