+++ frontend/pages/index.tsx
@@ -1,40 +1,25 @@
 import type { NextPage } from "next";
-import { useUser } from "@supabase/supabase-auth-helpers/react";
-import { supabaseClient } from "@supabase/supabase-auth-helpers/nextjs";
-import { forwardRef, useEffect, useRef, useState } from "react";
+import { useUser, useSupabaseClient } from "@supabase/auth-helpers-react";
+import { useEffect, useRef, useState } from "react";
 import "mapbox-gl/dist/mapbox-gl.css";
 import mapboxgl, { GeoJSONSource, Map } from "mapbox-gl";
 
-import { Auth } from "@supabase/ui";
 mapboxgl.accessToken = process.env.NEXT_PUBLIC_MAPBOX_TOKEN!;
-import type { Database } from "../src/types";
+import type { Database } from "../src/common/database-types";
 import { createFeatureCollection } from "../src/utils/create-feature-colection";
-type Tree = Omit<Database["public"]["Tables"]["trees"]["Row"], "geom"> & {
-	geom: GeoJSON.Geometry;
-};
-type District = Omit<
-	Database["public"]["Tables"]["districts"]["Row"],
-	"geom"
-> & { geom: GeoJSON.Geometry };
-
-type Adoption = Database["public"]["Tables"]["adoptions"]["Row"];
-type Watering = Database["public"]["Tables"]["waterings"]["Row"];
+import { AuthDetail } from "../components/AuthDetail";
+import { TreeInfo } from "../components/TreeInfo";
+import { AdoptonsDetail } from "../components/AdoptionsDetail";
+import { useGeojson } from "../src/lib/hooks/useGeojson";
+import { Seperator } from "../components/Seperator";
+import { WateringsDetail } from "../components/WateringsDetail";
+import { Adoption, Watering } from "../src/common/types";
+import { useDistricts } from "../src/lib/hooks/useDistricts";
 
 const Home: NextPage = () => {
+	const supabaseClient = useSupabaseClient<Database>();
 	const map = useRef<Map | null>(null);
-	async function initGeojson() {
-		const { data, error } = await supabaseClient
-			.from<Tree>("trees")
-			.select("*");
-		if (error) {
-			throw error;
-		}
-		const featureCollection = createFeatureCollection<Tree>(data);
 
-		setGeojson((_) => {
-			return featureCollection;
-		});
-	}
 	async function getFilteredTrees({
 		distance,
 		lat,
@@ -57,7 +42,8 @@ const Home: NextPage = () => {
 			console.error(error);
 			return;
 		}
-		const featureCollection = createFeatureCollection<Tree>(data);
+		//@ts-ignore
+		const featureCollection = createFeatureCollection(data);
 
 		setGeojson((_) => {
 			return featureCollection;
@@ -75,7 +61,8 @@ const Home: NextPage = () => {
 			console.error(error);
 			return;
 		}
-		const featureCollection = createFeatureCollection<Tree>(data);
+		// @ts-ignore
+		const featureCollection = createFeatureCollection(data);
 
 		setGeojson((_) => {
 			return featureCollection;
@@ -84,7 +71,7 @@ const Home: NextPage = () => {
 
 	async function getUserAdoptions(user_id: string) {
 		const { data, error } = await supabaseClient
-			.from<Adoption>("adoptions")
+			.from("adoptions")
 			.select("*")
 			.eq("user_id", user_id);
 		if (error) {
@@ -95,7 +82,7 @@ const Home: NextPage = () => {
 	}
 	async function getUserWaterings(user_id: string) {
 		const { data, error } = await supabaseClient
-			.from<Watering>("waterings")
+			.from("waterings")
 			.select("*")
 			.eq("user_id", user_id);
 		if (error) {
@@ -104,15 +91,12 @@ const Home: NextPage = () => {
 		}
 		setUserWaterings(data);
 	}
-	const { user, error: userError } = useUser();
-	const [geojson, setGeojson] = useState<
-		GeoJSON.FeatureCollection<GeoJSON.Geometry> | undefined
-	>(undefined);
+	const user = useUser();
+	const { geojson, setGeojson, initGeojson } = useGeojson(supabaseClient);
+	const districts = useDistricts(supabaseClient);
+
 	const [isGeolocating, setIsGeolocating] = useState<boolean>(false);
 	const [range, setRange] = useState<number>(100);
-	const [districts, setDistricts] = useState<
-		GeoJSON.FeatureCollection<GeoJSON.Geometry> | undefined
-	>(undefined);
 	const [currentLoction, setcurrentLoction] = useState<GeoJSON.Point>();
 	const [currentFeature, setcurrentFeature] =
 		useState<mapboxgl.MapboxGeoJSONFeature | null>(null);
@@ -207,24 +191,6 @@ const Home: NextPage = () => {
 	}, [user]);
 
 	useEffect(() => {
-		async function initDistricts() {
-			const { data, error } = await supabaseClient
-				.from<District>("districts")
-				.select("*");
-			if (error) {
-				throw error;
-			}
-			const featureCollection = createFeatureCollection<District>(data);
-			setDistricts((_) => {
-				return featureCollection;
-			});
-		}
-		initDistricts().catch(console.error);
-	}, []);
-	useEffect(() => {
-		initGeojson().catch(console.error);
-	}, []);
-	useEffect(() => {
 		if (!map) return;
 		if (!map.current) return;
 		if (!geojson) return;
@@ -257,38 +223,23 @@ const Home: NextPage = () => {
 		<>
 			<div className="container">
 				<div className="column">
-					<details>
-						<summary>{user ? "User info" : "Signup/login"}</summary>
-						<div className="row">
-							<div className="row-item" id="controlls">
-								{!user ? (
-									<Auth
-										// view="update_password"
-										supabaseClient={supabaseClient}
-										socialLayout="horizontal"
-										socialButtonSize="small"
-									/>
-								) : (
-									<>
-										<div>
-											<p>Email: {JSON.stringify(user.email, null, 2)}</p>
-											<p>id: {JSON.stringify(user.id, null, 2)}</p>
-											<button
-												onClick={() => {
-													supabaseClient.auth.signOut();
-													setUserAdoptions([]);
-													setUserWaterings([]);
-												}}
-											>
-												Sign out
-											</button>
-										</div>
-									</>
-								)}
-							</div>
-						</div>
-					</details>
+					{/*
+					----------------------------------------------------------------
+					The Auth detail section allowing the user to login
+					----------------------------------------------------------------
+					*/}
+					<AuthDetail
+						user={user}
+						supabaseClient={supabaseClient}
+						setUserAdoptions={setUserAdoptions}
+						setUserWaterings={setUserWaterings}
+					/>
 				</div>
+				{/*
+				----------------------------------------------------------------
+				The map
+				----------------------------------------------------------------
+				*/}
 				<div className="column">
 					<div className="row">
 						<div className="map row-item" id="map"></div>
@@ -339,114 +290,24 @@ const Home: NextPage = () => {
 							</div>
 						</div>
 					) : null}
+					{Seperator("Tree infos")}
 					<div className="row">
-						<div className="tree-info">
-							<div>
-								<strong>Tree id:</strong>{" "}
-								<span>
-									{currentFeature ? currentFeature?.properties?.id : ""}
-								</span>
-							</div>
-							<div>
-								<strong>Tree Type:</strong>{" "}
-								<span id="type">
-									{currentFeature ? currentFeature?.properties?.tree_type : ""}
-								</span>
-							</div>
-							<div>
-								<strong>Tree Age:</strong>{" "}
-								<span id="age">
-									{currentFeature ? currentFeature?.properties?.age : ""}
-								</span>
-							</div>
-							<div>
-								<strong>Tree Height:</strong>{" "}
-								<span id="height">
-									{currentFeature ? currentFeature?.properties?.height : ""}
-								</span>
-							</div>
-							<div>
-								<button
-									style={{ marginRight: "1em" }}
-									onClick={(e) => {
-										e.preventDefault();
-										async function adoptTree({
-											tree_id,
-											user_id,
-										}: {
-											tree_id: number;
-											user_id: string;
-										}) {
-											const { data, error } = await supabaseClient
-												.from<
-													Database["public"]["Tables"]["adoptions"]["Insert"]
-												>("adoptions")
-												.insert([{ tree_id, user_id }]);
-											if (error) {
-												console.error(error);
-												return;
-											}
-											console.log(data);
-											setUserAdoptions((prev) => {
-												return [...prev, data[0] as Required<Adoption>];
-											});
-										}
-										if (user) {
-											adoptTree({
-												tree_id: currentFeature?.properties?.id,
-												user_id: user?.id,
-											}).catch(console.error);
-										}
-									}}
-									disabled={!currentFeature}
-								>
-									Adopt
-								</button>
-
-								<button
-									onClick={() => {
-										async function waterTree({
-											tree_id,
-											amount,
-											user_id,
-										}: {
-											tree_id: number;
-											amount: number;
-											user_id: string;
-										}) {
-											const { data, error } = await supabaseClient
-												.from<
-													Database["public"]["Tables"]["waterings"]["Insert"]
-												>("waterings")
-												.insert([{ tree_id, amount, user_id }]);
-											if (error) {
-												console.error(error);
-												return;
-											}
-											console.log(data);
-											setUserWaterings((prev) => {
-												return [...prev, data[0] as Required<Watering>];
-											});
-										}
-										const amount = prompt(
-											"How much water do you want to add?",
-											"25",
-										);
-										if (user && amount) {
-											waterTree({
-												tree_id: currentFeature?.properties?.id,
-												amount: parseInt(amount),
-												user_id: user?.id,
-											}).catch(console.error);
-										}
-									}}
-									disabled={!currentFeature}
-								>
-									Water
-								</button>
-							</div>
-						</div>
+						{/*
+					----------------------------------------------------------------
+					The Tree info section Shows infrmation about the currently
+					selected tree
+					----------------------------------------------------------------
+					*/}
+						<TreeInfo
+							currentFeature={currentFeature}
+							user={user}
+							supabaseClient={supabaseClient}
+							setUserAdoptions={setUserAdoptions}
+							setUserWaterings={setUserWaterings}
+						/>
 					</div>
+					{Seperator("Filter trees by radius")}
+
 					<div className="row">
 						<div className="row-item">
 							<form
@@ -475,7 +336,7 @@ const Home: NextPage = () => {
 												lat: pos[0].geometry.coordinates[1],
 												//@ts-ignore
 												distance: parseInt(e.target.radius.value),
-											});
+											}).catch(console.error);
 										} else {
 											alert("Please select a tree");
 										}
@@ -493,6 +354,12 @@ const Home: NextPage = () => {
 							</form>
 						</div>
 					</div>
+					{Seperator("Filter trees by district")}
+					{/*
+					----------------------------------------------------------------
+					The district filter section allows selection of districs by dropdown selection. Filters using a remote procedure call
+					----------------------------------------------------------------
+					*/}
 					<div className="row">
 						<div className="row-item">
 							<form
@@ -525,109 +392,23 @@ const Home: NextPage = () => {
 							</form>
 						</div>
 					</div>
+					{Seperator("Waterings and Adoptions")}
 					<div className="row">
-						<div className="row-item">adopt + water input</div>
-					</div>
-					<div className="row">
-						<div className="row-item" id="" style={{ margin: "0" }}>
-							<h2>Waterings</h2>
-							{userWaterings.map((watering) => (
-								<div key={watering.id}>
-									<button
-										onClick={() => {
-											const pos = geojson?.features.filter(
-												(feature) =>
-													feature?.properties?.id === watering.tree_id,
-											);
-											if (pos && pos.length > 0) {
-												console.log("fly to");
-												map.current!.flyTo({
-													//@ts-ignore
-													center: pos[0].geometry.coordinates,
-													zoom: 17,
-												});
-											}
-										}}
-										style={{ marginRight: "1em" }}
-									>
-										🎯
-									</button>
-									<button
-										onClick={() => {
-											console.log("unwater");
-										}}
-										style={{ marginRight: "1em" }}
-									>
-										❌
-									</button>
-									<span>
-										Tree {watering.tree_id} got {watering.amount}l of water
-									</span>
-								</div>
-							))}
+						<div className="row-item">
+							<WateringsDetail
+								userWaterings={userWaterings}
+								map={map}
+								geojson={geojson}
+							/>
 						</div>
 						<div className="row-item" id="" style={{ margin: "0" }}>
-							<h2>Adoptions</h2>
-							{userAdoptions.map((adoption) => (
-								<div key={`${adoption.tree_id}:${adoption.user_id}`}>
-									<button
-										onClick={() => {
-											const pos = geojson?.features.filter(
-												(feature) =>
-													feature?.properties?.id === adoption.tree_id,
-											);
-											if (pos && pos.length > 0) {
-												console.log("fly to");
-												map.current!.flyTo({
-													//@ts-ignore
-													center: pos[0].geometry.coordinates,
-													zoom: 17,
-												});
-											}
-										}}
-										style={{ marginRight: "1em" }}
-									>
-										🎯
-									</button>
-									<button
-										style={{ marginRight: "1em" }}
-										onClick={() => {
-											async function removeAdoption({
-												user_id,
-												tree_id,
-											}: {
-												user_id: string;
-												tree_id: number;
-											}) {
-												const { data, error } = await supabaseClient
-													.from("adoptions")
-													.delete()
-													.eq("user_id", user_id)
-													.eq("tree_id", tree_id);
-												if (error) {
-													console.log(error);
-												}
-
-												console.log("unadopt", data);
-												setUserAdoptions((prev) =>
-													prev.filter(
-														(adoption) =>
-															adoption.user_id !== user_id &&
-															adoption.tree_id !== tree_id,
-													),
-												);
-											}
-											removeAdoption({
-												tree_id: adoption.tree_id,
-												user_id: adoption.user_id,
-											}).catch(console.error);
-										}}
-									>
-										❌
-									</button>
-									<span style={{ marginRight: "1em" }}>{adoption.tree_id}</span>
-								</div>
-							))}
+							<AdoptonsDetail
+								map={map}
+								geojson={geojson}
+								userAdoptions={userAdoptions}
+								supabaseClient={supabaseClient}
+								setUserAdoptions={setUserAdoptions}
+							/>
 						</div>
 					</div>
 				</div>
