+++ frontend/src/common/types.ts
@@ -1,0 +1,4 @@
+import { Database } from "./database-types";
+
+export type Adoption = Database["public"]["Tables"]["adoptions"]["Row"];
+export type Watering = Database["public"]["Tables"]["waterings"]["Row"];
